+++ hoomd/md/NeighborListGPUBinned.cc
@@ -127,7 +127,7 @@ void NeighborListGPUBinned::buildNlist(uint64_t timestep)
                                                            access_location::device,
                                                            access_mode::read);
 
-    ArrayHandle<unsigned int> d_head_list(m_head_list, access_location::device, access_mode::read);
+    ArrayHandle<size_t> d_head_list(m_head_list, access_location::device, access_mode::read);
     ArrayHandle<unsigned int> d_Nmax(m_Nmax, access_location::device, access_mode::read);
     ArrayHandle<unsigned int> d_conditions(m_conditions,
                                            access_location::device,
